Fix TODO about libapp.txt
Doesn't matter a ton what we name the download.  Giving it a generic name since we're going to change the format more over time (but it's not .txt).

Also moved more of the networking code into network.rs :)
